Fix signal/slot
svn path=/branches/work/konvi-kde4/; revision=930779

+++ src/quickconnectdialog.cpp
@@ -109,7 +109,7 @@ void QuickConnectDialog::slotOk()
 
         emit connectClicked(Konversation::PromptToReuseConnection,
                             hostNameInput->text().trimmed(),
-                            portInput->text(),
+                            portInput->text().toUInt(),
                             passwordInput->text(),
                             nickInput->text(),
                             "",
